validity-gadget-chain/x/validitygadgetchain/types/message_snapshot_proposal_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/twilight-project/validity-gadget-chain/testutil/sample"
)

func TestMsgSnapshotProposal_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSnapshotProposal
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSnapshotProposal{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSnapshotProposal{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
